Jenkinsfile: Updates for deb package build
The debian package build step needs to be run on a Ubuntu machine
as it requires the dpkg-deb utility which is not installed on redhat
lsf machines. Update the lsf bsub command accordingly for the overlay
build function. Also deploy the generated deb file as part of the
overlay deploy stage.

Signed-off-by: Christian Kohn <chris.kohn@amd.com>